feat: introduce specialized routers for typing
